Fix imports of moved objects
